+++ Source/CommandMap.h
@@ -83,7 +83,7 @@ public:
 // adds an entry to the message:command map, and a corresponding entry to the
 // command:message map will look up the string by the index (but it is preferred to
 // directly use the string)
-  void addCommandforMessage(unsigned int command, const MIDI_Message_ID& cc);
+  void addCommandforMessage(size_t command, const MIDI_Message_ID& cc);
 
   // adds an entry to the message:command map, and a corresponding entry to the
   // command:message map
