+++ Source/CommandMenu.cpp
@@ -54,7 +54,7 @@ void CommandMenu::setMsg(const MIDI_Message_ID& message) noexcept {
   message_ = message;
 }
 
-void CommandMenu::setSelectedItem(unsigned int index) {
+void CommandMenu::setSelectedItem(size_t index) {
   selected_item_ = index;
   if (index - 1 < LRCommandList::LRStringList.size())
     setButtonText(LRCommandList::LRStringList[index - 1]);
@@ -83,10 +83,10 @@ void CommandMenu::buttonClicked(juce::Button* /*button*/) {
       // disabling a previously mapped entry
 
       if (already_mapped)
-        subMenu.addColouredItem(index, command, juce::Colours::red, true,
+        subMenu.addColouredItem(static_cast<int>(index), command, juce::Colours::red, true,
           index == selected_item_);
       else
-        subMenu.addItem(index, command, true, index == selected_item_);
+        subMenu.addItem(static_cast<int>(index), command, true, index == selected_item_);
 
       index++;
     }
@@ -101,7 +101,7 @@ void CommandMenu::buttonClicked(juce::Button* /*button*/) {
   if ((result) && (command_map_)) {
       // user chose a different command, remove previous command mapping
       // associated to this menu
-    if (selected_item_ < std::numeric_limits<unsigned int>::max())
+    if (selected_item_ < std::numeric_limits<size_t>::max())
       command_map_->removeMessage(message_);
 
     if (result - 1 < LRCommandList::LRStringList.size())
