+++ Source/SendKeys.cpp
@@ -42,7 +42,7 @@ namespace {
   wchar_t MBtoWChar(const std::string& key) {
     wchar_t full_character;
     const auto return_value = MultiByteToWideChar(CP_UTF8, 0, key.data(),
-      key.size(), &full_character, 1);
+      static_cast<int>(key.size()), &full_character, 1);
     if (return_value == 0) {
       const auto er = GetLastError();
       if (er == ERROR_INVALID_FLAGS || er == ERROR_INVALID_PARAMETER)
@@ -227,7 +227,7 @@ void SendKeys::SendKeyDownUp(const std::string& key, const bool alt_opt,
   }
 
   //construct virtual keystroke sequence
-  std::vector<unsigned int> strokes{vk}; // start with actual key, then mods
+  std::vector<size_t> strokes{vk}; // start with actual key, then mods
   if (shift || (vk_modifiers & 0x1)) {
     strokes.push_back(VK_SHIFT);
   }
